[ParallelOps] Refactor kernel functions (#547)
* Combine

* Fix combine includes

* Partition + format

* Format and fix include

* Reduction

* Remove meta from reduction

* Replicate

* Fix reduction and replicate
