+++ include/flexflow/parallel_ops/partition.h
@@ -1,22 +1,13 @@
 #ifndef _FLEXFLOW_PARTITION_H
 #define _FLEXFLOW_PARTITION_H
 
-#include "flexflow/device.h"
-#include "flexflow/fftype.h"
 #include "flexflow/layer.h"
 #include "flexflow/node.h"
-#include "flexflow/op_meta.h"
 #include "flexflow/operator.h"
 #include "flexflow/parallel_ops/partition_params.h"
 #include "parallel_op.h"
 
 namespace FlexFlow {
-
-class RepartitionMeta : public OpMeta {
-public:
-  RepartitionMeta(FFHandler handle);
-  DataType data_type;
-};
 
 class Repartition : public ParallelOp {
 public:
@@ -63,13 +54,6 @@ public:
       std::vector<Legion::PhysicalRegion> const &regions,
       Legion::Context ctx,
       Legion::Runtime *runtime);
-  template <typename T>
-  static void
-      forward_kernel(T const *input_ptr, T *output_ptr, size_t num_elements);
-  template <typename T>
-  static void backward_kernel(T const *output_grad_ptr,
-                              T *input_grad_ptr,
-                              size_t num_elements);
   bool measure_operator_cost(Simulator *sim,
                              MachineView const &pc,
                              CostMetrics &cost_metrics) const override;
